+++ weblate/trans/models/conf.py
@@ -127,6 +127,7 @@ class WeblateConf(AppConf):
         'weblate.trans.checks.consistency.PluralsCheck',
         'weblate.trans.checks.consistency.SamePluralsCheck',
         'weblate.trans.checks.consistency.ConsistencyCheck',
+        'weblate.trans.checks.consistency.TranslatedCheck',
         'weblate.trans.checks.chars.NewlineCountingCheck',
         'weblate.trans.checks.markup.BBCodeCheck',
         'weblate.trans.checks.chars.ZeroWidthSpaceCheck',