+++ docs/changes.rst
@@ -7,6 +7,7 @@ weblate 2.13
 Released on ? 2017.
 
 * Fixed quality checks on translation templates.
+* Added quality check to trigger on losing translation.
 
 weblate 2.12
 ------------